Rework IFakeDependencyObjectHelpers.
